Move some functions to appropriate locations
